+++ .gitignore
@@ -3,7 +3,7 @@
 
 # python
 /venv
-/__pycache__
+__pycache__
 
 # topfarm & scipy output files
 *.out
