+++ lib/gwa_reader.py
